feat: add ioc container and font service for BlockSuite
