+++ packages/affine/services/src/font/index.ts
@@ -1,0 +1,5 @@
+import './scope.js';
+
+export { AffineCanvasTextFonts, CommunityCanvasTextFonts } from './consts.js';
+
+export type { FontConfig, FontService } from './interface.js';
