+++ packages/affine/services/src/font/interface.ts
@@ -1,0 +1,12 @@
+export interface FontConfig {
+  font: string;
+  weight: string;
+  url: string;
+  style: string;
+}
+
+export interface FontService {
+  ready: Promise<FontFace[]>;
+  clear: () => void;
+  load: (fonts: FontConfig[]) => void;
+}
