+++ packages/affine/services/src/index.ts
@@ -1,0 +1,5 @@
+import 'reflect-metadata';
+
+export * from './font/index.js';
+
+export { TYPES } from './types.js';
