+++ packages/affine/services/src/types.ts
@@ -1,0 +1,5 @@
+const TYPES = {
+  Font: Symbol.for('Font'),
+};
+
+export { TYPES };
