+++ packages/blocks/tsconfig.json
@@ -27,6 +27,9 @@
     },
     {
       "path": "../affine/components"
+    },
+    {
+      "path": "../affine/services"
     }
   ]
 }
