+++ packages/framework/block-std/package.json
@@ -23,9 +23,11 @@
     "@lit-labs/preact-signals": "^1.0.2",
     "@lit/context": "^1.1.2",
     "@types/hast": "^3.0.4",
+    "inversify": "^6.0.2",
     "lib0": "^0.2.95",
     "lit": "^3.1.4",
     "lz-string": "^1.5.0",
+    "reflect-metadata": "^0.2.2",
     "rehype-parse": "^9.0.0",
     "unified": "^11.0.5",
     "w3c-keyname": "^2.2.8",
